refactor: changing amount parameter to have decimal? data type instead of int?

+++ Platipus.Wallet.Api/Application/Requests/Wallets/Atlas/AtlasRefundRequest.cs
@@ -33,7 +33,7 @@ public sealed record AtlasRefundRequest(
                 request.Token,
                 request.RoundId,
                 request.TransactionId,
-                amount: (int)validAmount,
+                amount: validAmount,
                 clientId: request.ClientId,
                 cancellationToken: cancellationToken);
